New WebvizDialog Component (#275)
Dialog component draggable, with width and position adjustment when window width is resized. Functionality for handling multiple dialogs open at the same time, by controlling css for currently active dialog. This will make currently active dialog on top and with highlighting. The dialogs can be set modal - resulting in a clickable backdrop with adjusted opacity. When backdrop is clicked, the modal dialog is closed.